refactor: `useNftRanges` hook
