+++ src/components/[guild]/collect/components/CollectNft/components/AmountPicker.tsx
@@ -16,63 +16,12 @@ import useNftDetails from "components/[guild]/collect/hooks/useNftDetails"
 import Button from "components/common/Button"
 import FormErrorMessage from "components/common/FormErrorMessage"
 import useDebouncedState from "hooks/useDebouncedState"
-import { useEffect, useMemo, useState } from "react"
+import { useEffect, useState } from "react"
 import { useController, useFormContext } from "react-hook-form"
 import { Chains } from "wagmiConfig/chains"
 import { useCollectNftContext } from "../../CollectNftContext"
 import { CollectNftForm } from "../CollectNft"
-
-type MintingRange = {
-  name: string
-  icon: string
-  min: number
-  max: number
-}
-
-/**
- * Notes:
- *
- * - We'll always have 4 different ranges
- * - Infinite is when mintableAmountPerUser = 0, but we don't need to handle that case,
- *   since we won't show this component for that kind of NFTs
- */
-const getMintingRanges = (
-  mintableAmountPerUser: number
-): [MintingRange, MintingRange, MintingRange, MintingRange] => {
-  const cappedMintableAmountPerUser = Math.min(mintableAmountPerUser, 100)
-  const [range1Max, range2Max, range3Max] = [
-    Math.floor(cappedMintableAmountPerUser * 0.03),
-    Math.floor(cappedMintableAmountPerUser * 0.32),
-    Math.floor(cappedMintableAmountPerUser * 0.99),
-  ]
-
-  return [
-    {
-      name: "Shrimp",
-      icon: "🦐",
-      min: 1,
-      max: range1Max,
-    },
-    {
-      name: "Fish",
-      icon: "🐠",
-      min: range1Max + 1,
-      max: range2Max,
-    },
-    {
-      name: "Dolphin",
-      icon: "🐬",
-      min: range2Max + 1,
-      max: range3Max,
-    },
-    {
-      name: "Whale",
-      icon: "🐋",
-      min: range3Max + 1,
-      max: cappedMintableAmountPerUser,
-    },
-  ]
-}
+import useNftRanges from "../hooks/useNftRanges"
 
 const AmountPicker = () => {
   const { chain, nftAddress } = useCollectNftContext()
@@ -97,13 +46,7 @@ const AmountPicker = () => {
         )
       : 0
 
-  const ranges = useMemo(
-    () =>
-      mintableAmountPerUserFromContract
-        ? getMintingRanges(Number(mintableAmountPerUserFromContract))
-        : undefined,
-    [mintableAmountPerUserFromContract]
-  )
+  const ranges = useNftRanges()
 
   const rangeBgColor = useColorModeValue("gray.50", "blackAlpha.300")
   const circleBgColor = useColorModeValue("white", "gray.700")
